requirements.txt
aiohttp==3.6.2
arrow==0.15.6
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
discord-ext-menus==1.0.0a21+gd9965ec
discord.py==1.3.3
idna==2.9
importlib-metadata==1.6.0
jsonpickle==1.4.1
multidict==4.7.6
python-a2s==1.1.5
python-dateutil==2.8.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
zipp==3.1.0